requirements.txt
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html

autopep8==1.6.0
blis==0.7.7
catalogue==2.0.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==7.1.2
cycler==0.11.0
cymem==2.0.6
dill==0.3.5.1
docker-pycreds==0.4.0
fastai==2.6.3
fastcore==1.4.4
fastdownload==0.0.6
fastprogress==1.0.2
gitdb==4.0.9
GitPython==3.1.27
idna==2.10
importlib-metadata==4.11.4
Jinja2==2.11.3
joblib==1.1.0
kiwisolver==1.4.3
langcodes==3.3.0
loguru==0.6.0
MarkupSafe==2.0.1
matplotlib==3.2.2
murmurhash==1.0.7
numpy==1.21.6
opencv-python==4.1.2.30
packaging==21.3
pandas==1.3.5
pathlib==1.0.1
pathtools==0.1.2
pathy==0.6.1
pep8-naming==0.10.0
Pillow==7.1.2
preshed==3.0.6
prospector==1.7.7
protobuf==3.17.3
psutil==5.4.8
pydantic==1.8.2
pylint==2.13.5
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==3.13
requests==2.23.0
scikit-learn==1.0.2
scipy==1.4.1
sentry-sdk==1.6.0
setproctitle==1.2.3
shortuuid==1.0.9
six==1.15.0
smart-open==5.2.1
smmap==5.0.0
snowballstemmer==2.2.0
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
termcolor==1.1.0
thinc==8.0.17
threadpoolctl==3.1.0
tomli==2.0.1
torch @ https://download.pytorch.org/whl/cu113/torch-1.11.0%2Bcu113-cp37-cp37m-linux_x86_64.whl
torchvision @ https://download.pytorch.org/whl/cu113/torchvision-0.12.0%2Bcu113-cp37-cp37m-linux_x86_64.whl
tqdm==4.64.0
typer==0.4.1
typing-extensions==4.1.1
urllib3==1.24.3
wandb==0.12.20
wasabi==0.9.1
wrapt==1.14.1
yadisk==1.1.4
zipp==3.8.0